Added a script to turn the interaction features into fp16 format, allowing it to fit into ram of A100 gpu cluster, added suppor the data modules for lazy  loading and mmap loading for differnet ways to load into memory

+++ src/data/datamodule.py
@@ -535,7 +535,7 @@ class MaskedFormerTopsWsDataModule(LightningDataModule):
             src_mask = f["src_mask"][()]
 
             if self.load_interactions and "interactions" in f:
-                interactions = f["interactions"][()]
+                interactions = f["interactions"][()].astype(np.float16)
             else:
                 interactions = None
